+++ packages/libs/sdk-component-drivers/src/index.ts
@@ -19,3 +19,4 @@ export * from './DeviceListDriver';
 export * from './FlowButtonDriver';
 export * from './LastAuthBadgeDriver';
 export * from './TenantSelectorDriver';
+export * from './UserPasskeysDriver';